MIDI: Improve event timing accuracy
The stream generator now produces timestamps in seconds, double precison
floating point, and the player now operates with these timestamps. All
existing player modules have been adapted to the changes.

+++ Plugins/MIDI/MIDI/SCPlayer.h
@@ -17,7 +17,7 @@ class SCPlayer : public MIDIPlayer {
 	// close, unload
 	virtual ~SCPlayer();
 	
-	static uint32_t sampleRate();
+	static double sampleRate();
 
 	void setUrl(NSURL *url);
